refactor code to fix pattern and clean up
