+++ packages/@cdk-cosmos/core/src/galaxy/index.ts
@@ -1,0 +1,3 @@
+export { GalaxyCoreStack, GalaxyCoreStackProps } from './galaxy-core-stack';
+export { GalaxyCoreImport, GalaxyCoreImportProps } from './galaxy-core-import';
+export { GalaxyExtensionStack, GalaxyExtensionStackProps } from './galaxy-extension-stack';
