+++ packages/@cdk-cosmos/core/tsconfig.json
@@ -2,6 +2,7 @@
   "include": ["src"],
   "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
-    "outDir": "lib"
+    "outDir": "lib",
+    "declaration": false
   }
 }
